requirements.txt
discord.py
KahootPY
google_trans_new
pymongo
flask
dnspython
textblob
selenium
python-dotenv
thread6
nltk
passlib
SQLAlchemy
Werkzeug
gunicorn
gevent
ffmpeg
pynacl
youtube_dl==2021.12.17
